fix(schemas): validate regex safety at schema layer
